+++ src/collector/CollectorWrapper.ts
@@ -1,5 +1,5 @@
 import { createSessionStartedUserAction } from '../user-action/createSessionStartedUserAction'
-import { CollectorOptions, SessionStartedUserAction, SessionTraitValue } from '../types'
+import { CollectorOptions, CypressObject, SessionStartedUserAction, SessionTraitValue } from '../types'
 import UserActionHandler from '../user-action/UserActionHandler'
 import { debugSessionUserActionSender, defaultSessionUserActionSender } from '../user-action/sessionUserActionSender'
 import ISessionIdHandler from '../session-id-handler/ISessionIdHandler'
@@ -20,9 +20,12 @@ import { preventBadSessionTraitValue } from '../session-trait/checkSessionTraitV
 import { TargetEventListenerOptions } from '../event-listeners/TargetedEventListener'
 import createAsyncRequest from '../user-action/createAsyncRequest'
 import { trackingUrlStartPart } from '../gravityEndPoints'
+import CypressEventListener from '../event-listeners/CypressEventListener'
+import { IEventListener } from '../event-listeners/IEventListener'
+import IUserActionHandler from '../user-action/IUserActionHandler'
 
 class CollectorWrapper {
-  readonly userActionHandler: UserActionHandler
+  readonly userActionHandler: IUserActionHandler
   readonly userActionsHistory: MemoryUserActionsHistory
   readonly sessionTraitHandler: SessionTraitHandler
   readonly eventListenerHandler: EventListenersHandler
@@ -80,7 +83,7 @@ class CollectorWrapper {
       selectorsOptions: options.selectorsOptions,
     }
 
-    this.eventListenerHandler = new EventListenersHandler([
+    const eventListeners: IEventListener[] = [
       new ClickEventListener(this.userActionHandler, this.window, targetedEventListenerOptions),
       new KeyUpEventListener(this.userActionHandler, this.window, targetedEventListenerOptions),
       new KeyDownEventListener(
@@ -91,7 +94,13 @@ class CollectorWrapper {
       ),
       new ChangeEventListener(this.userActionHandler, this.window, targetedEventListenerOptions),
       new BeforeUnloadEventListener(this.userActionHandler, this.window),
-    ])
+    ]
+    const cypress = ((window as any).Cypress as CypressObject) ?? undefined
+    if (cypress !== undefined) {
+      eventListeners.push(new CypressEventListener(cypress, this.userActionHandler))
+    }
+
+    this.eventListenerHandler = new EventListenersHandler(eventListeners)
 
     this.trackingHandler.init(this.eventListenerHandler)