+++ src/event-listeners/ChangeEventListener.test.ts
@@ -1,17 +1,13 @@
 import { beforeEach, describe, expect, it, SpyInstance, vitest } from 'vitest'
-import UserActionHandler from '../user-action/UserActionHandler'
 import { fireEvent, getByRole, getByTestId, waitFor } from '@testing-library/dom'
 import userEvent from '@testing-library/user-event'
 import ChangeEventListener from '../event-listeners/ChangeEventListener'
-import { nop } from '../utils/nop'
 import createElementInJSDOM from '../test-utils/createElementInJSDOM'
-import MemorySessionIdHandler from '../session-id-handler/MemorySessionIdHandler'
 import * as createTargetedUserActionModule from '../user-action/createTargetedUserAction'
-import ISessionIdHandler from '../session-id-handler/ISessionIdHandler'
+import IUserActionHandler, { NopUserActionHandler } from '../user-action/IUserActionHandler'
 
 describe('ChangeEventListener', () => {
-  let sessionIdHandler: ISessionIdHandler
-  let userActionHandler: UserActionHandler
+  let userActionHandler: IUserActionHandler
   let handleSpy: SpyInstance
   let createTargetedUserActionSpy: SpyInstance
 
@@ -19,8 +15,7 @@ describe('ChangeEventListener', () => {
 
   beforeEach(() => {
     vitest.restoreAllMocks()
-    sessionIdHandler = new MemorySessionIdHandler(() => 'aaa-111', 600)
-    userActionHandler = new UserActionHandler(sessionIdHandler, 0, nop)
+    userActionHandler = new NopUserActionHandler()
     handleSpy = vitest.spyOn(userActionHandler, 'handle')
     createTargetedUserActionSpy = vitest.spyOn(createTargetedUserActionModule, 'createTargetedUserAction')
   })