+++ src/event-listeners/ChangeEventListener.ts
@@ -1,12 +1,12 @@
-import UserActionHandler from '../user-action/UserActionHandler'
 import { createTargetedUserAction } from '../user-action/createTargetedUserAction'
 import { HTMLInputWithValue, TargetedUserAction, UserActionType } from '../types'
 import { sanitizeHTMLElementValue } from '../utils/sanitizeHTMLElementValue'
 import TargetedEventListener, { TargetEventListenerOptions } from './TargetedEventListener'
 import { isTextField } from '../utils/listeners'
+import IUserActionHandler from '../user-action/IUserActionHandler'
 
 class ChangeEventListener extends TargetedEventListener {
-  constructor(userActionHandler: UserActionHandler, window: Window, options: TargetEventListenerOptions = {}) {
+  constructor(userActionHandler: IUserActionHandler, window: Window, options: TargetEventListenerOptions = {}) {
     super(userActionHandler, UserActionType.Change, window, options)
   }