+++ src/event-listeners/IEventListener.ts
@@ -1,0 +1,5 @@
+export interface IEventListener {
+  init: () => void
+
+  terminate: () => void
+}